Un-export internal-only symbols (METRIC/ASI prefixes, commas, fmtNum, sortedMedian, findBaselineResult, renderDashboardLines). Merge applyAutoresearchContractToExperimentState into contract.ts with re-export shim.
Result: {"status":"keep","check_ms":2850,"file_count":8,"line_count":2720,"biome_ms":506,"tsgo_ms":2344}
